Add copyright and checks
Add (C) to CERN for source files and Apache-2.0 license
in SPDX format. (C) for Resuse action is instead GPL3+.

Add Reuse action to ensure that license and (C) are always
checked.

Add AUTHORS file to ensure that all contributors are acknowledged.

All licensed used within the project are stored in the LICENSES
directory. Implement the Apache-2.0 license as a softlink to
the top level directory.