Added first version of concurrent BlockData container.
The block is templated on the data type, allocating the demanded number of elements in a pre-allocated buffer. The number of bytes to be allocated is returned by the static method BlockData<Type>::SizeOfInstance. The method SizeOfAlignAware should be used for allocating multiple blocks in the same buffer. The method NextElement gives write access to the next unused element, and the bracket operator can be used for read access.