Abort simulation if secondary cannot be produced (#64)
* examples: Add COPCORE_CUDA_CHECK to all API calls

This makes the program abort instead of going into a loop if the
kernel crashes, for example because of exhausted memory.

* Abort simulation if secondary cannot be produced

The assert() would only be present for Debug builds, leaving Release
builds with an "illegal memory access" due to accessing nullptr. Use
the existing COPCORE_EXCEPTION and extend it to print a message on
the device. This still isn't perfect because all threads will print,
but better than developers wondering what happened.