Correct number of secondaries for example1
number_of_secondaries is the total number of secondaries produced by
this track. To accumulate a global value, example1 must only add the
difference produced in the current step.
Also change the code in example2 to match the other two and always
add to the atomic (we need a more performant version of the scoring
anyway).