example6: Split relocation from navigation
When a particle leaves the current volume or enters a daughter, we
need to relocate to the next volume. This could involve walking the
hierarchy which is expensive, at least in the current LoopNavigator.
The problem is that if only one particle needs relocation, all others
have to wait.
As a first step, move the relocation step out of navigation, which
is called from the magnetic field, to a common place when a particle
is detected to need it. This already improves the total runtime from
around 1.4s to 1.3s on my system.