Initial skeleton for data import/export example
To decouple AdePT test programs from Geant4 as much as possible in
the testing phase, and to avoid potentially expensive recomputation
of physics data at application initialization, there is a use case to

- Persist geometry and physics data from a standard Geant4 application
  to file
- Read this file in a pure AdePT based application, reconsitituting
  the geometry and physics data for use on device

GDML/vecgeom is used for the geometry part, and G4HepEM for physics
data.

Provide a skeleton "example7" with placeholder applications for the
Geant4 (data export) and AdePT (data import) parts. Add build and
primary link options.