Create and use minimal interface target for AdePT
Add a CMake INTERFACE target for AdePT with main locations of headers
and dependency on CopCore as usage requirements. Dependency on VecGeom
via LoopNavigator deferred for now due to heavy CUDA dependence and need
to resolve a variety of usage requirements for CUDA/CPU variants.

Update tests and examples to link to the new target.