Add example with simplistic data structures
A SlotManager hands out monotonic slots (integers) as an index to a
plain array. This is up to 40% faster than example6. In the future,
fields could be split off into separate arrays if memory coalescing
proves important.

The disadvantage is that this scheme does not allow to reclaim memory
after a track was killed (which is a good thing if you want to see
where the tracks died after transportation finished). However, better
overall performance makes it feasible to just simulate the primaries
in multiple chunks and still be faster.