Fix (some) ODR violations in CopCore
Some function bodies are defined based on COPCORE_CUDA_COMPILER.
When linking compiled objects from both .cu and .cpp files (such
as the RaytraceBenchmark), and if the functions were not actually
inlined (for example in a Debug) build, the linker got to pick one
and the wrong implementation might be used.

This fixes #122 on my machine.