LoopNavigator: Correct handling of negative DistanceToIn
We never set the "last exited" volume, and there might not even be a
single volume in case of shared boundaries. So the existing check for
a negative distance was always false and the LoopNavigator returned
a zero-step in that case. This could make a track enter a daughter it
has just left. Any problems related to this would currently only be
hidden because tracks are pushed a tiny bit on every step, but this
is going to be reduced in the next commit.