example1[23]: Set cut value for default region
Also move construction of "DefaultRegionForTheWorld" before reading
of GDML to match G4RunManagerKernel, and call UpdateMaterialList.
This makes the list of couples (as printed by DumpCouples) identical
to a Geant4 application (including the indices), which is convenient
for debugging. No changes in the output for the two examples, probably
because no particle reaches a non-Air volume in the default region
where the energy thresholds change due to the cut value.