Continue tracking of strongly curling particles
Before, if a track did not arrive at its interaction point due to an
error in field propagation, it was simply killed on the basis that it
was likely a low-energetic electron. Instead keep it alive, apply the
continuous energy loss up to that intermediate point and enqueue the
particle for the next iteration step. In case it doesn't make progress,
the track will eventually be killed by the general infrastructure that
stops tracking if only electrons are left and no new secondary appeared
for 200 steps.

This approach also opens the door for optimizations such as limiting
the number of chord steps attempted in one tracking step, imposing a
maximum step limit, or only operating within the geometrical safety
sphere, if possible. However, this requires more thought since the
conversion of the partial (shorter) geometrical distance to a true
step length by MSC is only an approximation (but better than killing
the particle entirely). Also, experiments indicate that at least the
first approach, limiting the number of attempted chord steps, is not
effective or even counter-productive for more complex geometry setups
such as cms2018, so it is not attempted here.