Use safety in field propagation
Since the integration of MSC, we already compute the safety at the
beginnig of each charged step. Use it during field propagation to
skip querying the geometry if the remaining safety is larger than
the chord length.

The performance benefit depends on the combination of geometry and
strength of the magnetic field: From almost no change for TestEm3,
to around 10% with Bz=1T on cms2018 to 25% with Bz=3.8T.