TestEm3: add launch bounds to transportation kernels
This adds launch bounds with a safe configuration that still
allows a maximum of 256 registers per thread and 1 block per
SM. VecGeom requires 254 registers per thread right now, so
this is the highest number of threads that is safe to use
without needing to force VecGeom to use less registers at
compile time by adding, e.g. --maxrregcount 128 to the CUDA
compile options. If --maxrregcount 128 is used, then the number
of blocks can be increased to 2 for 256 threads (on the RTX 2070),
or a combination of 128 threads/4 blocks can be used.

There is a tradeoff between more threads/block and more blocks/SM for a
constant number of threads/SM. Using more threads/block reduces
scheduling overhead, and this is good for when many tracks are in
flight, as there are less blocks overall. Using more blocks/SM allows to
better distribute work across the GPU when there are very few tracks, as
a block is always scheduled on a single SM and smaller blocks can be
distributed more easily into more SMs.