Keep some Track members in registers
It is beneficial to keep some of the Track members in registers for the duration of the kernel.
The track variables are loaded at the beginning of the kernel and
written back when the track survives.
The signature of InitAsSecondary() had to be changed to not load from
the track again.

The results of the simulation change slighly because nvcc now generates
a few fma instructions in different places (thx to hahnjo for finding
out!).