Fix multiple events in integration examples (#230)
Commit 83836d06bc updated G4HepEm and adapted the RNG interface to
provide inline implementations instead of function pointers. The
intent was to use RANLUX++ on the GPU and continue using the CLHEP
generator (MIXMAX by default) on the CPU. However, the CUDA compiler
also generates host code for the inline function, which because of
symbol interception ends up being used by G4HepEm. This leads to
crashes because the inline function assumes to work on an instance
of RanluxppDouble and just writes to its members, leaving behind an
invalid state for CLHEP to seed at the beginning of the next event.

To solve this problem, surround the definition of the inline functions
with #ifdef __CUDA_ARCH__ to only make it visible to the compiler when
generating device code.