[async] Implement per-worker scoring on device.
Instead of one single scoring per AdePT, one scoring per G4 worker is
created. On the host, the scoring instances are stored in a vector,
whereas they are a malloc-ed array on the device. In this way, the
correct scoring instance can be accessed using the threadId of each
particle.

In order to have the scoring in a vector, the host side had to be made
copyable/moveable. The device side remains uninitialised when a host
instance is copied, though.