[async] Refactor AsyncExample to reduce the time between launches.
In the asynchronous transport loop, the exact population of particles on
the GPU doesn't need to be known on the host. Therefore, kernels can be
launched by just guessing the population of particles without having to
wait for device statistics to arrive. This reduces the idle time of the
GPU.