Force AdePT_G4_integration use as an RDC library
Link errors observed on CERN machines report missing symbols that
are in AdePT_G4_integration, and with this library on the link line.
The cause of this is not totally clear given success on other
platforms, though it is known that there is a circular dependency
between the AdePT and AdePT_G4_integration layer that might be an
underlying issue.

To attempt resolution/diagnosis, add a dummy .cu file to
AdePT_G4_integration. This forces it to be built as an RDC library and
thus onto the device link line alongside AdePT, potnetially resolving
the device link errors.