Explicitly create namespaced ALIAS targets for RDC libraries
If explicit ALIAS targets with the final install namespace (e.g.
"NS::Target" for "Target") are not created for an RDC library,
the installed targets have the wrong names set for the CUDA_RDC
properties. This results in both compile and (unseen) link errors
because the bare target names are interpreted as raw libraries,
not CMake targets. Dependencies are then not transmitted correctly,
resulting in compile/link issues for missing include paths etc.

Explicitly create ALIAS targets for CopCore and AdePT RDC libraries
to ensure they are installed with the correct install-time target
names for their `CUDA_RDC_...` properties.