The surface model uses BVH-accelerated Locate and Safety. No solids on GPU (#304)
This PR adapts the AdePT surface navigator to the index-based
interfaces. All navigation calls use the BVH surface navigator. The
solid model is no longer copied to GPU if `ADEPT_USE_SURF ` is ON.

The new interfaces seem to work correctly with non-overlapping
geometries, but we will need an example generating hits to validate
that.