Split AdePT kernels to separate physics from geometry (#313)
Adds the option to use a split version of the Electrons and Gammas
transport kernels. The main motivation for this is to split geometry
calls from physics as much as possible, to allow for better device
occupancy.

The gammas kernel is split in:
- First physics call to get the distance to the interaction
- Propagation
- Relocation
- Interactions

The electrons kernel is split in:
- First physics call to get the distance to interaction and MSC's
restricted geometrical length
- Propagation
- Second MSC call to recompute the physical length if needed
- Relocation
- Interactions

Splitting the kernels means that we now need to carry over some
information from one kernel to the next. Part of this information,
related to scoring and navigation, is stored as part of the AdePT track.
In order to keep track of the information needed for physics we add 3
arrays keeping a G4HepEM track of the corresponding type per AdePT
track.

This means that the memory needed per track increases:
- Single kernel:
AdePT Track: 264B

- Split kernels:
AdePT Track: 512B
HepEM Electron: 328B
HepEM Gamma: 176B

On average we need 789B per track when using split kernels