[mini] fix bug if next state was outside (#323)
If a step was recorded at the edge of the world such that the NextState
would be "outside" the reconstruction algorithm would try to access a
outside volume that doesn't exist in the lookup table. This would lead
to a segfault.

This PR adds a safeguard against that case. The scenario is unrealistic
in a real setup, but should be avoided for stability (and the error was
observed in some small tests)