reset buffer stats in HostScoring (#329)
This PR fixes a critical bug in the HostScoring implementation:

The 3 variables required for the next hit slot
```
  hostScoring_dev->fBufferStart        = 0;
  *(hostScoring_dev->fNextFreeHit_dev) = 0;
  *(hostScoring_dev->fUsedSlots_dev)   = 0;
```
were not reset between events. Although they were ever increased via `
unsigned int next = hostScoring_dev->fNextFreeHit_dev->fetch_add(1);`
(and the `fBufferStart` was set to be the `fNextHitFree`), it worked due
to the circular nature of the buffer `return next %
hostScoring_dev->fBufferCapacity;`. However, as soon as the maximum int
was exceeded, the simulation would simply stall and not do anything else
anymore. This PR fixes that by resetting the three variables between
events using a new function `ResetBufferStatsGPU`.

Note that there are a huge amount of changes due to the missing clang
formatting in the file. The only other change was to reduce some input
parameters for e.g., `CopyHitsToHost` and just use the hostscoring
itself to get the stats.