Update gamma kernel, add reproducibility CI test (#333)
This updates the gamma kernel, which is a leftover from #324.

The `SampleInteraction` only needs to be called if no boundary is hit
(as otherwise no physics process is used). This enables to keep the
`numIALeft` which is stored in the `currentTrack` and needs to be passed
to the G4HepEm theTrack. The previous early exit with setting the
winnerprocess to -1 caused non-reproducibility.

Since the code is reproducible with this fix, the CI test for
reproducibility with 8 ttbar events in CMS (no B field) is activated and
replaced the previous test that just ran one event.
Physics validation results are just as before.