ApplyCuts (rebase) (#334)
This PR adds the `ApplyCuts` used by G4 in AdePT. It is based and
supersedes PR #317. The `ApplyCuts` value is read from Geant4 and can be
toggled via `/process/em/applyCuts true`

Note that currently the positron cut is not yet available in G4HepEm
(but will be shortly), and we use just use the electron cut. This will
be fixed soon.

The restructuring of the scoring is now fully functional.

The run time shows that the `ApplyCuts` works as intended and
accelerates the simulation. Since the scoring was also effected, the
table also shows the results with `ApplyCuts = false`, which shows just
the new scoring implementation
| Setting | 4 ttbar 1 threads CMS (no Bfield) | 1e4 10 GeV electrons Em3
|
|-------------|-------------|---------------|
| master |  90.4422 s | 76.5938 s|
| new scoring | 90.3233 s| 75.4082 s |
| new scoring + ApplyCuts | 87.3564 s | 60.8005 s|

Additionally, the update of the time for gammas was incorrect (as it
would have been skipped if a boundary was hit), this is also fixed.

The physics validation looks really good using the `ApplyCuts` both in
G4 and AdePT (for 10^7 primaries

Co-authored-by: JuanGonzalezCaminero <j.gonzalez@cern.ch>