Better use of safety. (#322)
Implementing two safety features:

1. Safety caching: the tracks hold the point where safety is computed,
and the safety value in that point. When safety is needed in a different
point, the method track.GetSafety(new_point) will subtract from the
cached value and return the positive remainder or zero otherwise. The
GetSafety method can take an accurate limit value, and when the
remainder is smaller than this limit the returned safety is zero. It is
the user responsibility to call track.SetSafety(pos, safety) whenever
the safety is fully recomputed in a new position.

2. Safety calculation uses a new surface model feature only to compute
safety accurately when close to boundaries, and only use the aligned
bounding box safety when far away. The distance limit is passed as a
third parameter to `AdePTNavigator::ComputeSafety`, and must be
typically equal to the discrete interaction step.