Unify asynchronous and synchronous kernels (#338)
This PR unifies the synchronous and asynchronous kernels. 

The synchronous kernels included the usage of ApplyCuts and
TrackingCuts. Now, both the asynchronous mode and the synchronous mode
can use the same kernels for electrons and gammas. Note that the
previous asynchronous kernels used a split kernel approach for the
gammas along the interactions. To preserve those, everything related to
the split gamma kernel is just commented out. The performance between
both approaches was found to be practically the same (when not using
cuts in either of the kernels).

Some more cleaning should be done in the future, as the async mode uses
the global gVolAuxData, while the sync mode gets them passed as a
pointer (while still having the global one available, too!). This should
be cleaned together with the usage of the magnetic field as globals.

The physics results of the synchronous is as before and shows very good
agreement with G4 (< 2 per mille for 1e7 particles).

![Screenshot from 2025-01-29
10-05-42](https://github.com/user-attachments/assets/86628770-98f4-4c52-aca7-fa077e24a712)

ToDo before merge: 
- [ ] High statistics physics validation of the asynchronous mode