Improve async memory usage (#344)
This PR passes the debug levels properly to the async particle transport
again, now `adept/verbose` yields good debug printouts in the async
mode.

The handling of the GPUsteps is improved: instead of copying out the
full vector and making a shared pointer from each thread point to it
(this could cause issues if one worker was already done with its work,
not releasing the buffer), now just the steps that need to be processed
by that worker are copied to the queue of that worker.

Note that this is not the final implementation, since it is still too
slow and the GPU can easily run out of hitslots because the buffer is
not ready yet to be swapped back in, therefore a helper function is left
in the code and a few changes are not done to allow for flexibility to
test out the other approaches. Nonetheless, I already opened this PR
since the OOM crash when one worker finishes earlier than the rest might
be relevant right away.