Use direct linking to vecgeom's CUDA target to match AdePT
Reported by @SeverinDiederichs, trying to build AdePT with a static
VecGeom results in a failure of the `test_g4vg_link` buildtime test.
Likely caused by a corner case in `cuda_rdc_target_link_libraries`
not linking in everything from VecGeom in the right order, or failing
to device link in the static case. Linking of AdePT itself with
VecGeom and G4VG all fine.

Use direct call to target_link_libraries for the G4VG link test,
explicitly linking to both G4VG and the Vecgeomcuda targets. This gives
a correct link whether VecGeom was built with static or shared libs.