Storing the origin touchable of each track (#347)
This is part of the modifications required to run AdePT in ATLAS Athena.
With this PR we keep track of the origin touchable of each track, as
implemented
[here](https://github.com/mnovak42/g4hepem/pull/108/commits/80340477d82b82fda678edf67c7081e5c7ba88f3)
for the `G4HepEmTrackingManager`.

The main modifications needed for this are:
- Adding a new `NavigationState` field to AdePT tracks to store the
origin touchable
- Convert the track's origin touchable to a VecGeom `NavigationState`
when it goes into a GPU region
- When a primary track starts directly in a GPU region it's origin
touchable is not set, so we take the track's current position
- The origin is set for particles generated during GPU transport
- When particles leak back to G4, reconstruct the origin touchable and
add it to the track

It seems that in the case of Athena this information is needed for the
russian roulette. In the future we may also reconstruct and set the
touchable to the steps sent to scoring, but as this may affect
performance it should only be done if this information is actually
needed by the scoring code.

The result of a few short test runs doesn't show an impact on
performance from this addition alone. The tests were done with GPU
transport only in ECAL and HCAL, since the runs where we have leaks back
to G4 are the most affected:

`CMS2018, 32 TTbar, 8 Threads, Transport on ECAL+HCAL`

|        | Sync    | Async   |
|--------|---------|---------|
| Master | 293.599 | 279.3   |
| New    | 293.869 | 278.292 |