Improve handling of retrieved steps from the GPU in async mode (#350)
This PR improves the handling of the retrieved steps from the GPU in the
async mode.

Previously, the HitProcessingThread would always copy out the retrieved
steps to a vector and a unique_ptr would be pushed to the HitQueue. When
the G4 workers pop the item from the queue, the data would be released.
Unfortunately, the copying of GB of data could take significant time in
certain settings and block the buffer for longer than the GPU would
require to run out of HitSlots. This scenario would mostly appear in
testEm3 when shooting electrons, since every step of the full shower
would be recorded, leading to a vast amount of data to be copied.

This PR now changes the handling in the following way:

Now, there are still two buffers on device, but only one large circular
buffer on the CPU. Whenever the device buffer swaps, the retrieved steps
are copied into the circular host buffer. Then, the G4 workers are woken
up to directly score inside the circular buffer. When they finished the
scoring, the memory is released, i.e., the GPU can copy into it again.
If the circular host buffer is filled to more than 50% (which indicates
that the G4 workers are not scoring their hits, e.g., because they are
doing hadronics) the hits will be copied out before. If the G4 workers
are available, the copy can be prevented, and if the G4 workers are
busy, the hits can be copied out (e.g., as needed for CMS ttbar events).

This yields to greatly improved performance of the async mode for
applications where EM is largely dominating, such as testEm3.

The table shows the run time for T number of threads and E number of
events. The first rows denote testEm3 with 100 primary 10 GeV electrons
per event, the last two rows are ttbar events in CMS (without magnetic
field)

| Type    | Sync | Async master |  Async this PR |
|---------|------|--------------|-----|
| 1T 2E   | 1.9 | 1.27   |  1.2 |
| 2T 4E | 2.0 | 1.9 |  1.2 |
| 4T 16E | 4.16 | 5.0 |  2.6|
| 8T 32E | 4.6 | 10.1   | 2.8|
| 16T 64E | 6.6 |out of slots | 4.5 |
| 16T 256E | 24.1 | -  | 17.5 |
| 16T 1000E | 94.7 | - |70.2 |
| ttbar 8T 16E all sensitive | 79.6 | 72.4  | 70.3|
|ttbar 16T 64E | 134.126 | 118.461 | 115.5 |


This PR adds the async mode in the CI, now both are tested.
Also, this finally enabled physics validation in testEm3 with the async
AdePT, and it yields the same results as the sync AdePT:
<img width="583" alt="Screenshot 2025-02-22 at 17 06 47"
src="https://github.com/user-attachments/assets/0de4b38b-a5ab-4383-bd97-2e68c19642ce"
/>