Support G4PVReplica/G4VPVParameterised in geometry visitor
The latest updates to G4VG allow Geant4 parametrised/replica volumes
to be converted to a VecGeom representation by direct placements. This
means that there will be N VecGeom placements for each replicated
volume, N corresponding to the total number of replicas. Consequently,
the checks in visitGeometry will fail as:

1. The number of daughters in a given LV may be larger
2. Volume transformations may not match up as the original "parent"
   volume in Geant4 is compared against some copy of the replicated
   volume in VecGeom

Address the first issue by counting daughters on the Geant4 side as the
sum of the daughter multiplicities. For standard volumes, this gives
the same answer. Address the transformation problem by using the
concrete parameterization or G4ReplicaNavigation to transform the
Geant4 PV with the copy number of the compared VecGeom placed volume.