use a push in the solid model (#353)
The solid model relies on small pushes to relocate to the next volume.
If on boundary, by default it pushes by 1e-8, which is too small to help
some stuck particles. A push of 1e-8 can easily lead to no push at all
in one direction, which might be the direction where the push is needed.
This PR sets the push value to 1e-6 (`1e3 * kTolerance`) by default,
making it much more stable and preventing stuck particles that were
observed in the HGCAL.